Refactor primitive parsing
This commit refactors the way primitives are parsed, in preparation for
the subsequent implementation of primitive aliases (RFC #44). Before
this commit, primitive declarations in both structures and signatures
are parsed into the same `value_description` record as regular OCaml
values. As such, both are represented by the `Psig_value` variant in
signatures. After this commit, primitive declarations are parsed into a
separate `primitive_description` record, and are represented by a
new `Psig_primitive` variant in signatures. In structures,
`Pstr_primitive` is  updated to contain a `primitive_description` rather
than a `value_description`.

We also add a new `Pstr_val` node for the special case of `val v : t`
syntax in structures, which is supported for ppxes, but rejected by the
typechecker.

Finally, we propagate this split to the typed tree as well, resulting in
separate `Tsig_value of value_description` and
`Tsig_primitive of primitive_description` nodes.

There are some additional changes to tests which print out
representations of the AST.